refactor and rename classes
