+++ src/main/java/network/radicle/jetbrains/radiclejetbrainsplugin/config/RadicleSettingsView.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="network.radicle.jetbrains.radiclejetbrainsplugin.config.RadicleSettings">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="network.radicle.jetbrains.radiclejetbrainsplugin.config.RadicleSettingsView">
   <grid id="27dc6" binding="mainPanel" layout-manager="GridLayoutManager" row-count="3" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
@@ -18,7 +18,7 @@
           <component id="b8c65" class="javax.swing.JLabel">
             <constraints border-constraint="West"/>
             <properties>
-              <text value="Path to rad executable:"/>
+              <text value="Path to Rad executable:"/>
               <verticalAlignment value="0"/>
             </properties>
           </component>
