+++ src/main/resources/messages/RadicleBundle.properties
@@ -4,3 +4,4 @@ projectService=Project service: {0}
 selectExecutable=Select Executable
 radicle=Radicle
 radVersion=Rad version is : 
+radNotInstalled=Rad is not installed
